# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VisiteCRM API - SaaS CRM for tourism agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: pipelineStage
          in: query
          schema:
            type: ["string", "null"]
        - name: classification
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{id}/pipeline-stage:
    patch:
      operationId: updateClientPipelineStage
      tags: [clients]
      summary: Move client in pipeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stage:
                  type: string
              required: [stage]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # ─── NOTES ───────────────────────────────────────────────────────
  /clients/{clientId}/notes:
    get:
      operationId: listClientNotes
      tags: [notes]
      summary: List client notes
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createClientNote
      tags: [notes]
      summary: Create a note
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /clients/{clientId}/notes/{noteId}:
    delete:
      operationId: deleteClientNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── TRIPS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── DASHBOARD ─────────────────────────────────────────────────
...
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        phone:
          type: ["string", "null"]
        cpf:
          type: ["string", "null"]
        rg:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        addressCity:
          type: ["string", "null"]
        addressState:
          type: ["string", "null"]
        classification:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        pipelineStage:
          type: string
        totalSpent:
          type: number
        outstandingBalance:
          type: number
        npsScore:
          type: ["integer", "null"]
        observations:
          type: ["string", "null"]
        dreamDestinations:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        lastContactAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - email
        - whatsapp
        - classification
        - status
        - tags
        - pipelineStage
        - totalSpent
        - outstandingBalance
        - dreamDestinations
        - createdAt
        - updatedAt
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        phone:
          type: ["string", "null"]
        cpf:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        addressCity:
          type: ["string", "null"]
        addressState:
          type: ["string", "null"]
        observations:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        dreamDestinations:
          type: array
          items:
            type: string
      required: [name, email, whatsapp]
    UpdateClientBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        cpf:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        classification:
          type: ["string", "null"]
        tags:
          type: ["array", "null"]
          items:
            type: string
        observations:
          type: ["string", "null"]
        dreamDestinations:
          type: ["array", "null"]
          items:
            type: string
        addressCity:
          type: ["string", "null"]
        addressState:
          type: ["string", "null"]
        npsScore:
          type: ["integer", "null"]
        pipelineStage:
          type: ["string", "null"]
        lastContactAt:
          type: ["string", "null"]

    # ─── NOTE ──────────────────────────────────────────────────────
    Note:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        content:
          type: string
        isPrivate:
          type: boolean
        createdById:
          type: string
        createdAt:
          type: string
      required: [id, clientId, content, isPrivate, createdById, createdAt]
    CreateNoteBody:
      type: object
      properties:
        content:
          type: string
        isPrivate:
          type: boolean
      required: [content]

    # ─── TRIP ──────────────────────────────────────────────────────
...